se implemento recycleview para cursos y se arreglo el diseño
